+++ src/data/selectors.js
@@ -1,0 +1,8 @@
+import { createSelector } from 'reselect';
+
+export const app = state => state.app;
+
+export const gallery = createSelector(
+    app,
+    (app) => ({ projects : app })
+);